begun Refactoring AutoCVE Adapter
