+++ adapters/AutoCVE/AutoML.py
@@ -2,7 +2,8 @@ import json
 import os
 from JsonUtil import get_config_property
 
-from TabularDataAutoML import TabularDataAutoML
+from AutoCVEAdapter import AutoCVEAdapter
+
 
 if __name__ == '__main__':
     """
@@ -13,6 +14,6 @@ if __name__ == '__main__':
     with open(file_path) as file:
         process_json = json.load(file)
 
-    tabular_data_automl = TabularDataAutoML(process_json)
-    tabular_data_automl.execute_task()
+    tabular_data_automl = AutoCVEAdapter(process_json)
+    tabular_data_automl.start()
