+++ adapters/AutoCVE/Dockerfile
@@ -30,4 +30,4 @@ ENV RUNTIME=DOCKER
 ENV PYTHONUNBUFFERED=1
 ENV PYTHONPATH "AutoMLs:Utils"
 ENV PYTHON_ENV "python"
-ENTRYPOINT ["python", "Adapter_AutoCVE.py"]
+ENTRYPOINT ["python", "AutoCVEServer.py"]
