+++ adapters/AutoCVE/requirements.txt
@@ -1,5 +1,4 @@
-numpy==1.19.5
-grpcio==1.42.0
-protobuf==3.19.1
-jinja2==3.0.3
+grpcio
+dill
+jinja2
 git+https://github.com/celiolarcher/AUTOCVE@e87985d#egg=autocve
