+++ adapters/AutoGluon/requirements.txt
@@ -1,3 +1,4 @@
 autogluon
 grpcio
 dill
+jinja2
